Fixed simulator build

+++ Tests/Time/timer_wrapper_test.cpp
@@ -109,8 +109,8 @@ TEST_F(TimerWrapperTests, ConfigureTimer) {
 #define PERIOD 1000
     tim1.set_prescaler(PRESCALER_VAL);
     tim1.configure16bit(callback, 0, PERIOD);
-    EXPECT_EQ(TIM1_BASE->PSC, PRESCALER_VAL); /* set prescaler */
-    EXPECT_EQ(TIM1_BASE->ARR, PERIOD); /* set period */
+    EXPECT_EQ(static_cast<uint32_t>(TIM1_BASE->PSC), static_cast<uint32_t>(PRESCALER_VAL)); /* set prescaler */
+    EXPECT_EQ(static_cast<uint32_t>(TIM1_BASE->ARR), static_cast<uint32_t>(PERIOD)); /* set period */
     EXPECT_EQ(TIM1_BASE->CR1 & TIM_CR1_CEN, TIM_CR1_CEN); /* set counter enable */
 }
